:recycle: member -> img 패키지 이동
