+++ src/main/java/Idea/Archive/IdeaArchive/domain/member/entity/Member.java
@@ -58,5 +58,8 @@ public class Member {
     public void updateName(String name){
         this.name = name;
     }
+    public void updateProfileImg(String profileImageUrl) {
+        this.profileImageUrl = profileImageUrl;
+    }
 
 }
